fix(mcu): gear shifting + performance optimization
Fixes:
- Fixed JSON parsing: gear_ratios nested under 'transmission.gears'
- Added 200Nm default external_load for realistic RPM response
- Eliminated all double-precision math in audio ISR hot path
  (sample_player.cpp, engine_effects.cpp: double→float)
- Switched to -O3 -fsingle-precision-constant for MCU build
- Static ISR buffers (no more 4KB stack allocation per ISR call)
- Added DWT cycle counter for ISR CPU usage monitoring

Input:
- ButtonEvent library (from X-TRACK) with event callbacks
- ADC DMA mode for non-blocking throttle read
- millis()-based main loop scheduling (no blocking delay)

ISR budget: 3M cycles @ 288MHz/93.75Hz. Float-only path ~300K cycles (10%).

+++ core/include/core/engine_effects.h
@@ -49,7 +49,7 @@ private:
     size_t pulse_length_ = 0;
 
     // State
-    double phase_acc_ = 0.0; ///< Accumulator for firing interval
+    float phase_acc_ = 0.0f; ///< Accumulator for firing interval
     uint8_t current_cyl_ = 0; ///< Current cylinder index
     size_t pulse_pos_ = 0; ///< Current position in pulse playback
     bool pulse_active_ = false; ///< Is a pulse currently playing?
